add OwnedByAuthUser Rule in formRequest

+++ app/Http/Controllers/HouseController.php
@@ -62,6 +62,8 @@ class HouseController extends Controller
 
     public function destroy(House $house)
     {
+        $this->authorize('delete', $house);
+
         $house->delete();
         return redirect()->route('houses.index')->with('success', 'House deleted successfully!');
     }
